57/moves files around and fixes some small UI bugs.
