+++ Breeze.Web/src/components/dashboard/dataTables/ExpenseTable.tsx
@@ -1,22 +1,22 @@
-import * as React from 'react'
 import {
 	ColumnDef,
+	ColumnFiltersState,
 	SortingState,
 	VisibilityState,
 	flexRender,
 	getCoreRowModel,
-	getSortedRowModel,
 	getFilteredRowModel,
+	getSortedRowModel,
 	useReactTable,
-	ColumnFiltersState,
 } from '@tanstack/react-table'
+import dayjs from 'dayjs'
+import { ArrowUpDown } from 'lucide-react'
+import * as React from 'react'
 import { Button } from '../../../components/ui/button'
 import { Input } from '../../../components/ui/input'
-import { TableHeader, TableRow, TableHead, TableBody, TableCell, Table } from '../../../components/ui/table'
-import { ArrowUpDown } from 'lucide-react'
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../../components/ui/table'
 import { Expense } from '../../../services/hooks/expense/expenseServices'
 import { useBudgetContext } from '../../../services/providers/BudgetProvider'
-import dayjs from 'dayjs'
 import { ExpenseDialog } from '../dialogs/ExpenseDialog'
 
 // eslint-disable-next-line react-refresh/only-export-components
@@ -30,7 +30,7 @@ export const columns: ColumnDef<Expense>[] = [
 					onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
 				>
 					Name
-					<ArrowUpDown className='ml-2 h-4 w-4' />
+					<ArrowUpDown />
 				</Button>
 			)
 		},
@@ -44,7 +44,7 @@ export const columns: ColumnDef<Expense>[] = [
 					onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
 				>
 					Amount
-					<ArrowUpDown className='ml-2 h-4 w-4' />
+					<ArrowUpDown />
 				</Button>
 			)
 		},
@@ -62,7 +62,7 @@ export const columns: ColumnDef<Expense>[] = [
 					onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
 				>
 					Date
-					<ArrowUpDown className='ml-2 h-4 w-4' />
+					<ArrowUpDown />
 				</Button>
 			)
 		},
@@ -80,7 +80,7 @@ export const columns: ColumnDef<Expense>[] = [
 					onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
 				>
 					Category
-					<ArrowUpDown className='ml-2 h-4 w-4' />
+					<ArrowUpDown />
 				</Button>
 			)
 		},
