+++ Breeze.Web/src/components/deleteConfirmation/DeleteConfirmationDialog.tsx
