+++ Breeze.Web/src/components/goals/GoalDialog.tsx
@@ -3,7 +3,7 @@ import { Pencil } from 'lucide-react'
 import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { z } from 'zod'
-import { DeleteConfirmationDialog } from '../../components/reusable/DeleteConfirmationDialog'
+import { DeleteConfirmationDialog } from '../../components/deleteConfirmation/DeleteConfirmationDialog'
 import { Button } from '../../components/ui/button'
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '../../components/ui/dialog'
 import { FormField, FormItem, FormLabel, FormControl, Form } from '../../components/ui/form'
