+++ Breeze.Web/src/components/landingPage/LandingPage.tsx
@@ -1,7 +1,7 @@
 import { SignUpButton, useUser } from '@clerk/clerk-react'
-import { Button } from '../components/ui/button'
 import { Link } from 'react-router-dom'
-import AuthButton from '../components/auth/AuthButton'
+import { Button } from '../ui/button'
+import AuthButton from '../auth/AuthButton'
 
 export const LandingPage = () => {
 	const userIsSignedIn = useUser().isSignedIn
