+++ Breeze.Web/src/components/navigation/index.tsx
@@ -1,0 +1,12 @@
+import { MobileNavigation } from './MobileNavigation'
+import { DesktopNavigation } from './DesktopNavigation'
+
+export const Navigation = () => {
+	return (
+		<>
+			<DesktopNavigation />
+			<MobileNavigation />
+		</>
+	)
+}
+
